v1.0.0-trt-10.16

+++ cookbook/02-API/Layer/RotaryEmbedding/main.py
@@ -23,9 +23,9 @@ def case_simple():
     b, n, s, h = 1, 2, 4, 8
     max_position = 16
     data = {
-        "input": np.arange(b * n * s * h, dtype=np.float16).reshape(b, n, s, h) / 32,
-        "cos_cache": np.cos(np.linspace(0, 1, max_position * (h // 2), dtype=np.float32)).astype(np.float16).reshape(max_position, h // 2),
-        "sin_cache": np.sin(np.linspace(0, 1, max_position * (h // 2), dtype=np.float32)).astype(np.float16).reshape(max_position, h // 2),
+        "input": np.arange(b * n * s * h, dtype=np.float32).reshape(b, n, s, h) / 32,
+        "cos_cache": np.cos(np.linspace(0, 1, max_position * (h // 2), dtype=np.float32)).reshape(max_position, h // 2),
+        "sin_cache": np.sin(np.linspace(0, 1, max_position * (h // 2), dtype=np.float32)).reshape(max_position, h // 2),
         "position_ids": np.tile(np.arange(s, dtype=np.int64), (b, 1)),
     }
